requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
Django==4.0.6
idna==3.3
packaging==21.3
pgrep==2020.12.3
pyngrok==5.1.0
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
django-hitcount
django-crispy-forms
django-ckeditor

